refactor(state-data): move state data to the components/pages dir
move src/components/common/state-data.js -> src/components/pages/data/state-data.js
move src/scss/components/common/state-data.scss -> src/scss/components/pages/data/state-data.scss
