+++ src/components/pages/data/state-data.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import marked from 'marked'
 import { Link } from 'gatsby'
-import slug from '../../utilities/slug'
-import { UnstyledList } from './lists'
-import StateGrade from './state-grade'
-import SummaryTable from './summary-table'
-import '../../scss/components/common/state-data.scss'
+import slug from '../../../utilities/slug'
+import { UnstyledList } from '../../common/lists'
+import StateGrade from '../../common/state-grade'
+import SummaryTable from '../../common/summary-table'
+import '../../../scss/components/pages/data/state-data.scss'
 
 const State = ({ state }) => (
   <>
