+++ src/scss/components/pages/data/state-data.scss
@@ -1,4 +1,4 @@
-@import '../../breakpoints.scss';
+@import '../../../breakpoints.scss';
 
 .state-header {
   display: flex;
